Moved shared dtFabric code to helper mix-in
